Update to v1.12-pre.8

+++ Math/MapUtil.cs
@@ -4,14 +4,14 @@ namespace Vintagestory.API.MathTools
 {
     public class MapUtil
     {
-        public static long Index3dL(int x, int y, int z, int sizex, int sizez)
+        public static long Index3dL(int x, int y, int z, long sizex, long sizez)
         {
             return ((long)y * sizez + z) * sizex + x;
         }
 
-        public static int Index3d(int x, int y, int z, int sizex, int sizez)
+        public static int Index3d(int x, int y, int z, long sizex, long sizez)
         {
-            return (y * sizez + z) * sizex + x;
+            return (int)((y * sizez + z) * sizex + x);
         }
 
         public static int Index2d(int x, int y, int sizex)
@@ -25,18 +25,18 @@ namespace Vintagestory.API.MathTools
         }
 
 
-        public static void PosInt3d(int index, int sizex, int sizez, Vec3i ret)
+        public static void PosInt3d(int index, long sizex, long sizez, Vec3i ret)
         {
-            int x = index % sizex;
-            int y = index / (sizex * sizez);
-            int z = (index / sizex) % sizez;
+            long x = index % sizex;
+            long y = index / (sizex * sizez);
+            long z = (index / sizex) % sizez;
             
-            ret.X = x;
-            ret.Y = y;
-            ret.Z = z;
+            ret.X = (int)x;
+            ret.Y = (int)y;
+            ret.Z = (int)z;
         }
 
-        public static void PosInt3d(long index, int sizex, int sizez, Vec3i ret)
+        public static void PosInt3d(long index, long sizex, long sizez, Vec3i ret)
         {
             int x = (int)(index % sizex);
             int y = (int)(index / (sizex * sizez));
@@ -47,7 +47,7 @@ namespace Vintagestory.API.MathTools
             ret.Z = z;
         }
 
-        public static void PosInt2d(long index, int sizex, Vec2i ret)
+        public static void PosInt2d(long index, long sizex, Vec2i ret)
         {
             int x = (int)(index % sizex);
             int y = (int)(index / sizex);
